Fix JwtPayload exp data type Int to Long

+++ src/main/kotlin/cn/yiiguxing/plugin/translate/trans/edge/JwtPayload.kt
@@ -10,7 +10,7 @@ data class JwtPayload(
     @SerializedName("cognitive-services-endpoint")
     val cognitiveServicesEndpoint: String,
     @SerializedName("exp")
-    val exp: Int,
+    val exp: Long,
     @SerializedName("iss")
     val iss: String,
     @SerializedName("product-id")
